tests/wpt/tests/css/css-typed-om/the-stylepropertymap/properties/text-decoration-line.html

<!doctype html>
<meta charset="utf-8">
<title>'text-decoration-line' property</title>
<link rel="help" href="https://drafts.css-houdini.org/css-typed-om-1/#dom-stylepropertymap-get">
<link rel="help" href="https://drafts.css-houdini.org/css-typed-om-1/#dom-stylepropertymap-set">
<link rel="help" href="https://drafts.css-houdini.org/css-typed-om-1/#property-stle-value-normalization">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="../../resources/testhelper.js"></script>
<script src="resources/testsuite.js"></script>
<body>
<div id="log"></div>
<script>
'use strict';

runPropertyTests('text-decoration-line', [
  { syntax: 'none' },
  { syntax: 'underline' },
  { syntax: 'overline' },
  { syntax: 'line-through' },
  { syntax: 'blink' },
  { syntax: 'spelling-error' },
  { syntax: 'grammar-error' },
]);

</script>
