third_party/blink/web_tests/external/wpt/css/css-typed-om/the-stylepropertymap/properties/text-decoration-line-expected.txt

This is a testharness.js-based test.
[FAIL] Can set 'text-decoration-line' to the 'blink' keyword: blink
  assert_equals: expected "blink" but got "none"
Harness: the test ran to completion.
